+++ include/WestBot/clientstuff.h
@@ -45,8 +45,8 @@ public slots:
 signals:
     void statusChanged(bool);
     void hasReadSome(SimData msg);
-    void updateRobotPos( uint8_t id, uint8_t color, int x, int y, int theta );
-    void updatePuckPos( uint8_t id, uint8_t color, int x, int y, bool onTable = true );
+    void updateRobotPos( uint8_t id, uint8_t color, double x, double y, double theta );
+    void updatePuckPos( uint8_t id, uint8_t color, double x, double y, bool onTable = true );
 
 private slots:
     void readyRead();
